requirements.txt
azure-ai-language-questionanswering==1.1.0
azure-ai-textanalytics==5.3.0
openai==1.59.6
python-dotenv==1.0.1
streamlit==1.41.1
azure-ai-language-conversations==1.0.0